Move topology diff to render package.
