+++ report/topology.go
@@ -3,7 +3,6 @@ package report
 import (
 	"fmt"
 	"net"
-	"reflect"
 	"strings"
 )
 
@@ -118,47 +117,6 @@ func netsContain(nets []*net.IPNet, ip net.IP) bool {
 	}
 	return false
 }
-
-// Diff is returned by TopoDiff. It represents the changes between two
-// RenderableNode maps.
-type Diff struct {
-	Add    []RenderableNode `json:"add"`
-	Update []RenderableNode `json:"update"`
-	Remove []string         `json:"remove"`
-}
-
-// TopoDiff gives you the diff to get from A to B.
-func TopoDiff(a, b RenderableNodes) Diff {
-	diff := Diff{}
-
-	notSeen := map[string]struct{}{}
-	for k := range a {
-		notSeen[k] = struct{}{}
-	}
-
-	for k, node := range b {
-		if _, ok := a[k]; !ok {
-			diff.Add = append(diff.Add, node)
-		} else if !reflect.DeepEqual(node, a[k]) {
-			diff.Update = append(diff.Update, node)
-		}
-		delete(notSeen, k)
-	}
-
-	// leftover keys
-	for k := range notSeen {
-		diff.Remove = append(diff.Remove, k)
-	}
-
-	return diff
-}
-
-// ByID is a sort interface for a RenderableNode slice.
-type ByID []RenderableNode
-
-func (r ByID) Len() int           { return len(r) }
-func (r ByID) Swap(i, j int)      { r[i], r[j] = r[j], r[i] }
-func (r ByID) Less(i, j int) bool { return r[i].ID < r[j].ID }
 
 // Validate checks the topology for various inconsistencies.
 func (t Topology) Validate() error {
